+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 
 
+## 4.3.1.3
+### Fixed
+- design but in webpage (would load webpage in class constructor)
+- search engine updated to latest Webpage
+
 ## 4.3.1.2
 ### Added
 - url validator with regex